fix: save new tunnels to config, add name editing, SSH failure alerts
- Fix config.yml not saving on close after adding a tunnel: removed
  premature self.data update; DeepDiff now correctly detects new tunnels
- Add Name field to TunnelConfig (edit form) with rename support;
  accepted signal updates the tunnel display name in the main window
- Validate name/port conflicts in both add and edit dialogs (empty name,
  duplicate name, duplicate port all block saving with a clear error)
- Add QProcess error/finish signal handlers: show warning dialog and
  reset the start/stop button when SSH fails to start or exits non-zero
- Fix new tunnel connect issue: _stopping flag prevents false alerts on
  deliberate stops; process errors are now surfaced instead of silent

+++ app.py
@@ -38,14 +38,15 @@ def parse_ssh_config():
     return hosts
 
 class TunnelConfig(QDialog):
-    def __init__(self, parent, data):
+    def __init__(self, parent, name, data):
         super(TunnelConfig, self).__init__(parent)
-        
+
         self.ui = Ui_TunnelConfig()
         self.ui.setupUi(self)
-        
+
+        self.ui.name_field.setText(name)
         self.ui.remote_address.setText(data.get(KEYS.REMOTE_ADDRESS))
-        self.ui.browser_open.setText(data.get(KEYS.BROWSER_OPEN))
+        self.ui.browser_open.setText(data.get(KEYS.BROWSER_OPEN) or "")
         self.ui.local_port.setValue(data.get(KEYS.LOCAL_PORT))
 
         self.ui.proxy_host.addItems(parse_ssh_config())
@@ -59,23 +60,34 @@ class TunnelConfig(QDialog):
         self.ui.copy.clicked.connect(self.do_copy_ssh_command)
 
         self.used_ports = set()
+        self.used_names = set()
 
         self.render_ssh_command()
 
     def accept(self):
+        name = self.get_name()
+        if not name:
+            QMessageBox.warning(self, LANG.OOPS, LANG.NAME_EMPTY)
+            return
+        if name in self.used_names:
+            QMessageBox.warning(self, LANG.OOPS, LANG.NAME_IN_USE.format(name))
+            return
         if self.ui.local_port.value() in self.used_ports:
             QMessageBox.warning(self, LANG.OOPS, LANG.PORT_IN_USE.format(self.ui.local_port.value()))
             return
         super().accept()
 
+    def get_name(self):
+        return self.ui.name_field.text().strip()
+
     def render_ssh_command(self):
         ssh_command = F"ssh -L 127.0.0.1:{self.ui.local_port.value()}:{self.ui.remote_address.text()} {self.ui.proxy_host.currentText()}"
         self.ui.ssh_command.setText(ssh_command)
-        
+
     def do_copy_ssh_command(self):
         cb = QApplication.clipboard()
         cb.setText(self.ui.ssh_command.text())
-        
+
     def as_dict(self):
         return {
             KEYS.REMOTE_ADDRESS: self.ui.remote_address.text(),
@@ -153,10 +165,18 @@ class AddTunnelDialog(QDialog):
         self.local_port.valueChanged.connect(self.render_ssh_command)
 
         self.used_ports = set()
+        self.used_names = set()
 
         self.render_ssh_command()
 
     def accept(self):
+        name = self.get_name()
+        if not name:
+            QMessageBox.warning(self, LANG.OOPS, LANG.NAME_EMPTY)
+            return
+        if name in self.used_names:
+            QMessageBox.warning(self, LANG.OOPS, LANG.NAME_IN_USE.format(name))
+            return
         if self.local_port.value() in self.used_ports:
             QMessageBox.warning(self, LANG.OOPS, LANG.PORT_IN_USE.format(self.local_port.value()))
             return
@@ -183,27 +203,32 @@ class AddTunnelDialog(QDialog):
 class Tunnel(QWidget):
     def __init__(self, name, data):
         super(Tunnel, self).__init__()
-        
+
         self.ui = Ui_Tunnel()
         self.ui.setupUi(self)
-        
-        self.tunnelconfig = TunnelConfig(self, data)
+
+        self.tunnelconfig = TunnelConfig(self, name, data)
         self.tunnelconfig.setWindowTitle(name)
         self.tunnelconfig.setModal(True)
+        self.tunnelconfig.accepted.connect(self._on_config_accepted)
         self.ui.name.setText(name)
-        
+
         self.tunnelconfig.icon = F"./icons/{name}.png"
-        
+
         if not os.path.exists(self.tunnelconfig.icon):
           self.tunnelconfig.icon = ICONS.TUNNEL
-        
+
         self.ui.icon.setPixmap(QPixmap(self.tunnelconfig.icon))
         self.ui.action_tunnel.clicked.connect(self.do_tunnel)
         self.ui.action_settings.clicked.connect(self.tunnelconfig.show)
         self.ui.action_open.clicked.connect(self.do_open_browser)
-        
+
         self.process = None
-        
+        self._stopping = False
+
+    def _on_config_accepted(self):
+        self.ui.name.setText(self.tunnelconfig.get_name())
+
     def do_open_browser(self):
         browser_open = self.tunnelconfig.ui.browser_open.text()
         if browser_open:
@@ -211,42 +236,64 @@ class Tunnel(QWidget):
             local_port = self.tunnelconfig.ui.local_port.value()
             new_url = urlobj._replace(netloc=F"{urlobj.hostname}:{local_port}").geturl()
             QDesktopServices.openUrl(QUrl(new_url))
-        
+
     def do_tunnel(self):
         if self.process:
             self.stop_tunnel()
         else:
             self.start_tunnel()
-    
+
     def start_tunnel(self):
+        self._stopping = False
         params = self.tunnelconfig.ui.ssh_command.text().split(" ")
-        
+
         self.process = QProcess()
+        self.process.errorOccurred.connect(self._on_process_error)
+        self.process.finished.connect(self._on_process_finished)
         self.process.start(params[0], params[1:])
-                    
+
         self.ui.action_tunnel.setIcon(QIcon(ICONS.STOP))
-        
+
         self.do_open_browser()
-    
+
     def stop_tunnel(self):
-        try:
-            self.process.kill()
+        self._stopping = True
+        if self.process:
+            try:
+                self.process.kill()
+            except:
+                pass
             self.process = None
-        except:
-            pass
-        
         self.ui.action_tunnel.setIcon(QIcon(ICONS.START))
-        
+
+    def _on_process_error(self, error):
+        if not self._stopping:
+            self._stopping = True
+            name = self.ui.name.text()
+            QMessageBox.warning(self, LANG.OOPS, LANG.SSH_ERROR.format(name))
+        self._cleanup_process()
+
+    def _on_process_finished(self, exit_code, exit_status):
+        if not self._stopping and exit_code != 0:
+            self._stopping = True
+            name = self.ui.name.text()
+            QMessageBox.warning(self, LANG.OOPS, LANG.SSH_EXITED.format(name, exit_code))
+        self._cleanup_process()
+
+    def _cleanup_process(self):
+        self.process = None
+        self.ui.action_tunnel.setIcon(QIcon(ICONS.START))
+
 class TunnelManager(QWidget):
     def __init__(self):
         super().__init__()
-        
+
         with open(CONF_FILE, "r") as fp:
             self.data = yaml.load(fp, Loader=yaml.FullLoader)
 
         self.grid = QGridLayout(self)
         self.tunnels = []
-        
+
         for i, name in enumerate(sorted(self.data.keys())):
             tunnel = Tunnel(name, self.data[name])
             self.tunnels.append(tunnel)
@@ -255,7 +302,7 @@ class TunnelManager(QWidget):
             tunnel.ui.action_settings.clicked.connect(
                 lambda checked=False, t=tunnel: self.show_tunnel_settings(t)
             )
-        
+
         self.add_button = QPushButton(LANG.ADD_TUNNEL)
         self.add_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
         self.add_button.clicked.connect(self.do_add_tunnel)
@@ -267,40 +314,33 @@ class TunnelManager(QWidget):
 
         self.grid.addWidget(self.add_button, i+1, 0)
         self.grid.addWidget(self.kill_button, i+2, 0)
-        
+
         self.setLayout(self.grid)
         self.resize(10, 10)
         self.setWindowTitle(LANG.TITLE)
         self.setWindowIcon(QIcon(ICONS.TUNNEL))
         self.show()
-    
+
     def _get_used_ports(self, exclude=None):
         return {t.tunnelconfig.ui.local_port.value() for t in self.tunnels if t is not exclude}
 
+    def _get_used_names(self, exclude=None):
+        return {t.ui.name.text() for t in self.tunnels if t is not exclude}
+
     def show_tunnel_settings(self, tunnel):
         tunnel.tunnelconfig.used_ports = self._get_used_ports(exclude=tunnel)
+        tunnel.tunnelconfig.used_names = self._get_used_names(exclude=tunnel)
         tunnel.tunnelconfig.show()
 
     def do_add_tunnel(self):
         dialog = AddTunnelDialog(self)
         dialog.used_ports = self._get_used_ports()
+        dialog.used_names = self._get_used_names()
         if dialog.exec() != QDialog.DialogCode.Accepted:
             return
 
         name = dialog.get_name()
-        if not name:
-            return
-
-        if name in self.data:
-            mb = QMessageBox(self)
-            mb.setIcon(QMessageBox.Icon.Warning)
-            mb.setText(F"Tunnel '{name}' already exists.")
-            mb.setWindowTitle(LANG.OOPS)
-            mb.exec()
-            return
-
         data = dialog.as_dict()
-        self.data[name] = data
 
         tunnel = Tunnel(name, data)
         self.tunnels.append(tunnel)
@@ -325,15 +365,15 @@ class TunnelManager(QWidget):
             os.system(CMDS.SSH_KILL_WIN)
         else:
             os.system(CMDS.SSH_KILL_NIX)
-            
+
     def closeEvent(self, event):
         data = {}
         for tunnel in self.tunnels:
             name = tunnel.ui.name.text()
             data[name] = tunnel.tunnelconfig.as_dict()
-        
+
         changed = DeepDiff(self.data, data, ignore_order=True)
-        
+
         if changed:
             timestamp = int(time.time())
             shutil.copy(CONF_FILE, F"{CONF_FILE}-{timestamp}")
@@ -344,12 +384,12 @@ class TunnelManager(QWidget):
                 for config in sorted(backup_configs, reverse=True)[10:]:
                     os.remove(config)
         event.accept()
-    
+
 if __name__ == '__main__':
     app = QApplication(sys.argv)
-    
+
     sm = QSharedMemory("3866273d-f4d5-4bf3-b27b-772ca7915a61")
-    
+
     if not sm.create(1):
         mb = QMessageBox()
         mb.setIcon(QMessageBox.Icon.Information)
@@ -364,7 +404,7 @@ if __name__ == '__main__':
         mb.setWindowTitle(LANG.OOPS)
         mb.setStandardButtons(QMessageBox.StandardButton.Close)
         mb.show()
-    else:        
+    else:
         tm = TunnelManager()
-            
+
     sys.exit(app.exec())
